Componentes del dashboard

+++ MegoPoderes/src/app/pages/_layout/components/aside/aside.component.ts
@@ -13,20 +13,20 @@ import { SidebarService } from 'src/app/_theme/core/services/sidebar.service';
   styleUrls: ['./aside.component.scss'],
 })
 export class AsideComponent implements OnInit {
+
   usuario: Usuario;
   funcionalidadPadre: Funcionalidades;
-  perId: number;
-  disableAsideSelfDisplay: boolean;
   headerLogo: string;
   brandSkin: string;
-  ulCSSClasses: string;
   location: Location;
-  asideMenuHTMLAttributes: any = {};
-  asideMenuCSSClasses: string;
-  asideMenuDropdown;
+  asideSelfMinimizeToggle = true;
   brandClasses: string;
   asideMenuScroll = 1;
-  asideSelfMinimizeToggle = false;
+  ulCSSClasses: string;
+  asideMenuCSSClasses: string;
+  // disableAsideSelfDisplay: boolean;
+  // asideMenuHTMLAttributes: any = {};
+  // asideMenuDropdown;
 
   constructor(
     private layout: LayoutService,
@@ -35,30 +35,31 @@ export class AsideComponent implements OnInit {
     private usuarioService: LoginService,
     private route: Router
   ) {
-    this.usuario = this.usuarioService.usuario;
-    this.perId = this.usuario.id_perfil;
-    this.obtenerMenu(this.perId);
+
   }
 
   ngOnInit(): void {
-
+    this.usuario = this.usuarioService.usuario;
+    if (this.usuario) { this.obtenerMenu(this.usuario.id_perfil); } else {   this.route.navigate(['/auth/login']); }
     this.headerLogo = this.getLogo();
-    // load view settings
-    this.disableAsideSelfDisplay = this.layout.getProp('aside.self.display') === false;
-    this.brandSkin = this.layout.getProp('brand.self.theme');
-    this.ulCSSClasses = this.layout.getProp('aside_menu_nav');
-    this.asideMenuCSSClasses = this.layout.getStringCSSClasses('aside_menu');
-    this.asideMenuHTMLAttributes = this.layout.getHTMLAttributes('aside_menu');
-    this.asideMenuDropdown = this.layout.getProp('aside.menu.dropdown') ? '1' : '0';
-    this.brandClasses = this.layout.getProp('brand');
     this.asideSelfMinimizeToggle = this.layout.getProp(
       'aside.self.minimize.toggle'
     );
+    this.brandClasses = this.layout.getProp('brand');
+    this.ulCSSClasses = this.layout.getProp('aside_menu_nav');
     this.asideMenuScroll = this.layout.getProp('aside.menu.scroll') ? 1 : 0;
+    this.asideMenuCSSClasses = this.layout.getStringCSSClasses('aside_menu');
+
+    // load view settings
+    // this.disableAsideSelfDisplay = this.layout.getProp('aside.self.display') === false;
+    // this.brandSkin = this.layout.getProp('brand.self.theme');
+    // this.asideMenuHTMLAttributes = this.layout.getHTMLAttributes('aside_menu');
+    // this.asideMenuDropdown = this.layout.getProp('aside.menu.dropdown') ? '1' : '0';
     // this.asideMenuCSSClasses = `${this.asideMenuCSSClasses} ${this.asideMenuScroll === 1 ? 'scroll my-4 ps ps--active-y' : ''}`;
     // Routing
     this.location = this.loc;
-    // console.log(this.location);
+    const angularRoute = this.loc.path();
+    const url = window.location.href;
   }
 
   private getLogo() {
@@ -69,32 +70,35 @@ export class AsideComponent implements OnInit {
     }
   }
 
-  obtenerMenu( perfilId: number): void {
+  async obtenerMenu(perfilId: number) {
 
-    this.sidebarService.get_funcionalidadesPadre(perfilId)
-      .then((resp: any) => {
+    if (perfilId) {
+      await this.sidebarService.get_funcionalidadesPadre(perfilId)
+        .then((resp: any) => {
 
-        this.funcionalidadPadre = resp;
+          this.funcionalidadPadre = resp;
 
-        localStorage.setItem( 'menu', JSON.stringify(this.funcionalidadPadre));
+          if (resp.length === 0) {
 
-        if (resp.length === 0) {
+            localStorage.clear();
+            this.route.navigate(['/auth/login']);
 
-          localStorage.clear();
+          } else {
+            // localStorage.setItem( 'menu', JSON.stringify(resp));
+            this.route.navigate(['/inicio']);
+            // console.log(this.funcionalidadPadre);
+          }
+
+        }).catch(err => {
+
+          alert('Error al obtener funcionalidades asignadas');
           this.route.navigate(['/auth/login']);
+          console.warn(err);
 
-        }else{
+        });
 
-          this.route.navigate(['/dashboard']);
-
-        }
-
-      }).catch(err => {
-
-        alert('Error al obtener funcionalidades asignadas');
-        this.route.navigate(['/auth/login']);
-        console.warn(err);
-
-      });
+    } else {
+      this.route.navigate(['/auth/login']);
+     }
   }
 }
